Move RichHistogram into "rich" submodule.
